+++ razorpay/constants/device.py
@@ -1,0 +1,8 @@
+"""POS Device related constants."""
+
+
+class DeviceMode:
+    """Possible types of device mode."""
+
+    WIRED = "wired"
+    WIRELESS = "wireless"