+++ razorpay/constants/http_status_code.py
@@ -1,0 +1,9 @@
+"""HTTP Response status codes."""
+
+
+class HttpStatusCode:
+    """HTTP Status codes."""
+
+    OK = 200
+    NO_CONTENT = 204
+    REDIRECT = 300